refactor DrivesDao into DriveDao
