+++ lib/models/daos/daos.dart
@@ -1,3 +1,2 @@
 export 'drive_dao/drive_dao.dart';
-export 'drives_dao.dart';
 export 'profile_dao.dart';
