+++ lib/models/daos/drive_dao/create_drive_result.dart
@@ -1,0 +1,9 @@
+part of 'drive_dao.dart';
+
+class CreateDriveResult {
+  final String driveId;
+  final String rootFolderId;
+  final SecretKey driveKey;
+
+  CreateDriveResult(this.driveId, this.rootFolderId, this.driveKey);
+}
