+++ lib/models/daos/drive_dao/drive_dao.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:ardrive/entities/entities.dart';
 import 'package:ardrive/services/services.dart';
+import 'package:arweave/arweave.dart';
 import 'package:cryptography/cryptography.dart';
 import 'package:equatable/equatable.dart';
 import 'package:moor/moor.dart';
@@ -10,6 +11,7 @@ import 'package:uuid/uuid.dart';
 
 import '../../database/database.dart';
 
+part 'create_drive_result.dart';
 part 'drive_dao.g.dart';
 part 'drive_order.dart';
 part 'folder_node.dart';
@@ -22,6 +24,125 @@ class DriveDao extends DatabaseAccessor<Database> with _$DriveDaoMixin {
   final _uuid = Uuid();
 
   DriveDao(Database db) : super(db);
+
+  /// Creates a drive with its accompanying root folder.
+  Future<CreateDriveResult> createDrive({
+    @required String name,
+    @required String ownerAddress,
+    @required String privacy,
+    Wallet wallet,
+    String password,
+    SecretKey profileKey,
+  }) async {
+    final driveId = _uuid.v4();
+    final rootFolderId = _uuid.v4();
+
+    var insertDriveOp = DrivesCompanion.insert(
+      id: driveId,
+      name: name,
+      ownerAddress: ownerAddress,
+      rootFolderId: rootFolderId,
+      privacy: privacy,
+    );
+
+    SecretKey driveKey;
+    if (privacy == DrivePrivacy.private) {
+      driveKey = await deriveDriveKey(wallet, driveId, password);
+      insertDriveOp = await _addDriveKeyToDriveCompanion(
+          insertDriveOp, profileKey, driveKey);
+    }
+
+    await batch((batch) {
+      batch.insert(drives, insertDriveOp);
+
+      batch.insert(
+        folderEntries,
+        FolderEntriesCompanion.insert(
+          id: rootFolderId,
+          driveId: driveId,
+          name: name,
+          path: '',
+        ),
+      );
+    });
+
+    return CreateDriveResult(
+      driveId,
+      rootFolderId,
+      driveKey,
+    );
+  }
+
+  /// Adds or updates the user's drives with the provided drive entities.
+  Future<void> updateUserDrives(
+          Map<DriveEntity, SecretKey> driveEntities, SecretKey profileKey) =>
+      db.batch((b) async {
+        for (final entry in driveEntities.entries) {
+          final entity = entry.key;
+
+          var driveCompanion = DrivesCompanion.insert(
+            id: entity.id,
+            name: entity.name,
+            ownerAddress: entity.ownerAddress,
+            rootFolderId: entity.rootFolderId,
+            privacy: entity.privacy,
+            dateCreated: Value(entity.commitTime),
+            lastUpdated: Value(entity.commitTime),
+          );
+
+          if (entity.privacy == DrivePrivacy.private) {
+            driveCompanion = await _addDriveKeyToDriveCompanion(
+                driveCompanion, profileKey, entry.value);
+          }
+
+          b.insert(
+            drives,
+            driveCompanion,
+            onConflict:
+                DoUpdate((_) => driveCompanion.copyWith(dateCreated: null)),
+          );
+        }
+      });
+
+  Future<void> writeDriveEntity({
+    String name,
+    DriveEntity entity,
+  }) {
+    assert(entity.privacy == DrivePrivacy.public);
+
+    final companion = DrivesCompanion.insert(
+      id: entity.id,
+      name: name,
+      ownerAddress: entity.ownerAddress,
+      rootFolderId: entity.rootFolderId,
+      privacy: entity.privacy,
+      dateCreated: Value(entity.commitTime),
+      lastUpdated: Value(entity.commitTime),
+    );
+
+    return into(drives).insert(
+      companion,
+      onConflict: DoUpdate((_) => companion.copyWith(dateCreated: null)),
+    );
+  }
+
+  Future<DrivesCompanion> _addDriveKeyToDriveCompanion(
+    DrivesCompanion drive,
+    SecretKey profileKey,
+    SecretKey driveKey,
+  ) async {
+    final iv = Nonce.randomBytes(96 ~/ 8);
+    final encryptedWallet = await aesGcm.encrypt(
+      await driveKey.extract(),
+      secretKey: profileKey,
+      nonce: iv,
+    );
+
+    return drive.copyWith(
+      encryptedKey: Value(encryptedWallet),
+      keyEncryptionIv: Value(iv.bytes),
+    );
+  }
 
   /// Returns the encryption key for the specified drive.
   ///
