+++ test/blocs/drive_create_cubit_test.dart
@@ -13,7 +13,7 @@ import '../utils/utils.dart';
 void main() {
   group('DriveCreateCubit', () {
     Database db;
-    DrivesDao drivesDao;
+    DriveDao driveDao;
 
     ArweaveService arweave;
     DrivesCubit drivesCubit;
@@ -24,7 +24,7 @@ void main() {
 
     setUp(() {
       db = getTestDb();
-      drivesDao = db.drivesDao;
+      driveDao = db.driveDao;
 
       arweave = ArweaveService(Arweave());
       drivesCubit = MockDrivesCubit();
@@ -40,7 +40,7 @@ void main() {
 
       driveCreateCubit = DriveCreateCubit(
         arweave: arweave,
-        drivesDao: drivesDao,
+        driveDao: driveDao,
         drivesCubit: drivesCubit,
         profileCubit: profileCubit,
       );
